moved state machine reads out of state_machine_objects.rs so as not to acquire a lock to get the reads
